GH #225
